requirements.txt
mysql-connector-python
django-rest-swagger